Convert person object in a data.frame.

Results in a data.frame with the given name, family name, e-mail, ORCID, affiliation and role. Missing elements result in an empty string (""). Persons with multiple roles will have the roles as a comma separated string.

Usage

author2df(person)

Arguments

person

The person object or a list of person objects, NA or NULL. Any "character" is converted to a person object using as.person() with a warning.

See also