Make sure that the required environment variables are set on GitHub

Some actions will fail when these environment variables are not set. This function does only work on GitHub.

check_environment(x = ".")

Arguments

x

Either a checklist object or a path to the source code. Defaults to ..

Value

An invisible checklist object.

See also