Check the packages for linters

This functions does static code analysis. It relies on lintr::lint_package(). We recommend that you activate all code diagnostics in RStudio to help meeting the requirements. You can find this in the menu Tools > Global options > Code > Diagnostics. Please have a look at vignette("philosophy") for more details on the rules.

Usage

check_lintr(x = ".", quiet = FALSE)

Arguments

x

Either a checklist object or a path to the source code. Defaults to ..

quiet

Whether to print check output during checking.

Details

When check_lintr() runs on a git repository, it checks whether the organisation has a custom .lintr file in their checklist repository. If so, it uses this file. Otherwise it looks for a local .lintr file in the project directory. If this file is not present, it uses the default .lintr file provided with the checklist package.

See also