Standardised test for an R source repository

A convenience function that runs test on a project with only .R and .Rmd files. The details section lists the relevant functions. When you fixed a problem, you can speed things up by running only the related check. We still recommend to run check_source() before you push to GitHub. And only push when the functions indicate that there are no problems. This catches most problems before sending the code to GitHub.

check_source(x = ".", fail = !interactive())

Arguments

x

Either a checklist object or a path to the source code. Defaults to ..

fail

Should the function return an error in case of a problem? Defaults to TRUE on non-interactive session and FALSE on an interactive session.

Details

List of checks in order:

check_lintr()

check_filename()

See also