A function that asks a yes or no question to the user

A function that asks a yes or no question to the user

yesno(...)

Arguments

...

Currently ignored

Value

A logical where TRUE implies a "yes" answer from the user.

See also

Author

Hadley Wickham Hadley@Rstudio.com Largely based on devtools:::yesno(). The user gets three options in an random order: 2 for "no", 1 for "yes". The wording for "yes" and "no" is random as well. This forces the user to carefully read the question.