The checklist R6 class

A class which contains all checklist results.

Public fields

package

A logical indicating whether the source code refers to a package.

Active bindings

get_checked

A vector with checked topics.

get_path

The path to the package.

fail

A logical indicating if all checks passed.

template

A list for a check list template.

Methods

Public methods

checklist$add_error()

checklist$add_linter()

checklist$add_motivation()

checklist$add_notes()

checklist$add_rcmdcheck()

checklist$add_warnings()

checklist$allowed()

checklist$confirm_motivation()

checklist$new()

checklist$print()

checklist$clone()

Method add_error()

Add errors

Usage

checklist$add_error(errors, item)

Arguments

errors

A vector with errors.

item

The item on which to store the errors.

Method add_linter()

Add results from lintr::lint_package()

Usage

checklist$add_linter(linter)

Arguments

linter

A vector with linter errors.

Method add_motivation()

Add motivation for allowed issues.

Usage

checklist$add_motivation(which = c("warnings", "notes"))

Arguments

which

Which kind of issue to add.

Method add_notes()

Add notes

Usage

checklist$add_notes(notes)

Arguments

notes

A vector with notes.

Method add_rcmdcheck()

Add results from rcmdcheck::rcmdcheck

Usage

checklist$add_rcmdcheck(errors, warnings, notes)

Arguments

errors

A vector with errors.

warnings

A vector with warning messages.

notes

A vector with notes.

Method add_warnings()

Add warnings

Usage

checklist$add_warnings(warnings)

Arguments

warnings

A vector with warnings.

Method allowed()

Add allowed warnings and notes

Usage

checklist$allowed(
  warnings = vector(mode = "list", length = 0),
  notes = vector(mode = "list", length = 0)
)

Arguments

warnings

A vector with allowed warning messages. Defaults to an empty list.

notes

A vector with allowed notes. Defaults to an empty list.

package

Does the check list refers to a package. Defaults to TRUE.

Method confirm_motivation()

Confirm the current motivation for allowed issues.

Usage

checklist$confirm_motivation(which = c("warnings", "notes"))

Arguments

which

Which kind of issue to confirm.

Method new()

Initialize a new Checklist object.

Usage

checklist$new(x)

Arguments

x

The path to the root of the package.

Method print()

Print the Checklist object.

Usage

checklist$print(...)

Arguments

...

currently ignored.

Method clone()

The objects of this class are cloneable with this method.

Usage

checklist$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.